+++ src/Cocorico/UserBundle/Controller/Dashboard/ProfileContactController.php
@@ -12,9 +12,6 @@
 
 namespace Cocorico\UserBundle\Controller\Dashboard;
 
-use Cocorico\UserBundle\Entity\UserAddress;
-use Cocorico\UserBundle\Event\UserEvent;
-use Cocorico\UserBundle\Event\UserEvents;
 use Cocorico\UserBundle\Form\Type\ProfileContactFormType;
 use FOS\UserBundle\Model\UserInterface;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Method;
@@ -40,33 +37,37 @@ class ProfileContactController extends Controller
      *
      * @return \Symfony\Component\HttpFoundation\Response
      */
-    public function editContactAction(Request $request)
+    public function editAction(Request $request)
     {
         $user = $this->getUser();
         if (!is_object($user) || !$user instanceof UserInterface) {
             throw new AccessDeniedException('This user does not have access to this section.');
         }
 
-        $form = $this->createEditContactForm($user);
-        $form->handleRequest($request);
+        $formHandler = $this->get('cocorico_user.form.handler.contact');
+        $user = $formHandler->init($user);
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            //Eventually change user profile data before update it
-            $event = new UserEvent($user);
-            $this->get('event_dispatcher')->dispatch(UserEvents::USER_PROFILE_UPDATE, $event);
-            $user = $event->getUser();
+        $form = $this->createContactForm($user);
+        $success = $formHandler->process($form);
 
-            $this->get("cocorico_user.user_manager")->updateUser($user);
+        $session = $this->container->get('session');
+        $translator = $this->container->get('translator');
 
-            $this->container->get('session')->getFlashBag()->add(
+        if ($success > 0) {
+            $session->getFlashBag()->add(
                 'success',
-                $this->container->get('translator')->trans('user.edit.contact.success', array(), 'cocorico_user')
+                $translator->trans('user.edit.contact.success', array(), 'cocorico_user')
             );
 
             return $this->redirect(
                 $this->generateUrl(
                     'cocorico_user_dashboard_profile_edit_contact'
                 )
+            );
+        } elseif ($success < 0) {
+            $session->getFlashBag()->add(
+                'error',
+                $translator->trans('user.edit.contact.error', array(), 'cocorico_user')
             );
         }
 
@@ -86,12 +87,8 @@ class ProfileContactController extends Controller
      *
      * @return \Symfony\Component\Form\Form The form
      */
-    private function createEditContactForm($user)
+    private function createContactForm($user)
     {
-        $addresses = $user->getAddresses();
-        if (count($addresses) == 0) {
-            $user->addAddress(new UserAddress());
-        }
         $form = $this->get('form.factory')->createNamed(
             'user',
             new ProfileContactFormType(),
@@ -101,7 +98,6 @@ class ProfileContactController extends Controller
                 'action' => $this->generateUrl('cocorico_user_dashboard_profile_edit_contact'),
             )
         );
-
 
         return $form;
     }